03-WindLeft implemented from IREC code, if/else logic for servo actuation questionable second look good idea
