+++ Polaris/src/states/02-Freefall.cpp
@@ -1,6 +1,6 @@
 #include "State.h"
 #include "02-Freefall.h"
-#include "WindLeft.h"
+#include "03-WindLeft.h"
 #include "FlightParams.hpp"
 
 Freefall::Freefall(FlashChip *flash, AttitudeStateEstimator *attitudeStateEstimator, XbeeProSX *xbee, struct Servos *servos, OpenMV *openMV) :  State(flash, attitudeStateEstimator, xbee, servos, openMV){}
