--- Polaris/src/states/WindLeft.cpp
@@ -1,38 +1,0 @@
-#include "WindLeft.h"
-#include "State.h"
-#include "04-HoldLeft.h"
-#include "FlightParams.hpp"
-
-WindLeft::WindLeft(FlashChip *flash, AttitudeStateEstimator *attitudeStateEstimator, XbeeProSX *xbee, struct Servos *servos, OpenMV *openMV) :  State(flash, attitudeStateEstimator, xbee, servos, openMV){}
-
-void WindLeft::initialize_impl() {
-	this->stateStartTime = this->currentTime;
-}
-
-void WindLeft::loop_impl() {
-	this->stateTime = this->currentTime - this->stateStartTime; 
-	if(stateTime < 1500){
-		this->servos->paraServo_3->writeServo(SERVO_COUNTER_WIND); 
-		this->servos->paraServo_4->writeServo(SERVO_CLOCK_WIND); ; //Check Servo values 
-	}
-	else{
-		this->servos->paraServo_3->writeServo(SERVO_CENTER); 
-		this->servos->paraServo_4->writeServo(SERVO_CENTER); ; //Check Servo values
-	}
-	
-}
-
-State *WindLeft::nextState_impl() {
-	if (this->stateTime > MAX_SERVO_WIND_TIME /*||servo values = MAX_SERVO_POS*/)
-	{
-		//stop servos 
-		Serial.println("Entering HoldLeft!"); 
-		return new HoldLeft(this->flash, this->attitudeStateEstimator, this->xbee, this->servos, this->openMV);
-	}
-	return nullptr;
-}
-
-enum StateId WindLeft::getId()
-{
-    return StateId::ID_WindLeft;
-}
